server/requirements.txt
fastapi
python-socketio
uvicorn
websockets
motor